--- a/libdebug/architectures/aarch64/aarch64_thread_context.py
+++ b/libdebug/architectures/aarch64/aarch64_thread_context.py
@@ -14,14 +14,20 @@ if TYPE_CHECKING:
     from libdebug.architectures.aarch64.aarch64_ptrace_register_holder import (
         Aarch64PtraceRegisterHolder,
     )
+    from libdebug.debugger.internal_debugger import InternalDebugger
 
 
 class Aarch64ThreadContext(ThreadContext):
     """This object represents a thread in the context of the target aarch64 process. It holds information about the thread's state, registers and stack."""
 
-    def __init__(self: Aarch64ThreadContext, thread_id: int, registers: Aarch64PtraceRegisterHolder) -> None:
+    def __init__(
+        self: Aarch64ThreadContext,
+        thread_id: int,
+        registers: Aarch64PtraceRegisterHolder,
+        internal_debugger: InternalDebugger,
+    ) -> None:
         """Initialize the thread context with the given thread id."""
-        super().__init__(thread_id, registers)
+        super().__init__(thread_id, registers, internal_debugger)
 
         # Register the thread properties
         self._register_holder.apply_on_thread(self, Aarch64ThreadContext)