--- a/libdebug/data/breakpoint.py
+++ b/libdebug/data/breakpoint.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from libdebug.state.thread_context import ThreadContext
 
 
-@dataclass
+@dataclass(eq=False)
 class Breakpoint:
     """A breakpoint in the target process.
 
@@ -39,13 +39,11 @@ class Breakpoint:
     callback: None | Callable[[ThreadContext, Breakpoint], None] = None
     condition: str = "x"
     length: int = 1
-    _enabled: bool = True
 
-    _linked_thread_ids: list[int] = field(default_factory=list)
-    # The thread ID that hit the breakpoint
-
-    _disabled_for_step: bool = False
-    _changed: bool = False
+    _enabled: bool = field(default=True, init=False, repr=False)
+    _linked_thread_ids: list[int] = field(default_factory=list, init=False, repr=False)
+    _disabled_for_step: bool = field(default=False, init=False, repr=False)
+    _changed: bool = field(default=False, init=False, repr=False)
 
     @property
     def enabled(self: Breakpoint) -> bool:
@@ -77,11 +75,3 @@ class Breakpoint:
             return False
 
         return internal_debugger.resume_context.event_hit_ref.get(thread_context.thread_id) == self
-
-    def __hash__(self: Breakpoint) -> int:
-        """Hash the breakpoint object by its memory address, so that it can be used in sets and dicts correctly."""
-        return hash(id(self))
-
-    def __eq__(self: Breakpoint, other: object) -> bool:
-        """Check if two breakpoints are equal."""
-        return id(self) == id(other)