--- a/libdebug/data/breakpoint.py
+++ b/libdebug/data/breakpoint.py
@@ -9,11 +9,10 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
 
-from libdebug.debugger.internal_debugger_instance_manager import provide_internal_debugger
-
 if TYPE_CHECKING:
     from collections.abc import Callable
 
+    from libdebug.debugger.internal_debugger import InternalDebugger
     from libdebug.state.thread_context import ThreadContext
 
 
@@ -44,17 +43,18 @@ class Breakpoint:
     _linked_thread_ids: list[int] = field(default_factory=list, init=False, repr=False)
     _disabled_for_step: bool = field(default=False, init=False, repr=False)
     _changed: bool = field(default=False, init=False, repr=False)
+    _internal_debugger: InternalDebugger = field(default=None, init=True, repr=False)
 
     @property
     def enabled(self: Breakpoint) -> bool:
         """Whether the breakpoint is enabled or not."""
-        provide_internal_debugger(self)._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         return self._enabled
 
     @enabled.setter
     def enabled(self: Breakpoint, value: bool) -> None:
         """Set the enabled state of the breakpoint."""
-        provide_internal_debugger(self)._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         self._enabled = value
         self._changed = True
 
@@ -68,10 +68,9 @@ class Breakpoint:
 
     def hit_on(self: Breakpoint, thread_context: ThreadContext) -> bool:
         """Returns whether the breakpoint has been hit on the given thread context."""
-        internal_debugger = provide_internal_debugger(self)
-        internal_debugger._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
 
         if not self._enabled:
             return False
 
-        return internal_debugger.resume_context.event_hit_ref.get(thread_context.thread_id) == self
+        return self._internal_debugger.resume_context.event_hit_ref.get(thread_context.thread_id) == self