--- a/libdebug/data/memory_map_list.py
+++ b/libdebug/data/memory_map_list.py
@@ -6,18 +6,25 @@
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from libdebug.data.memory_map import MemoryMap
-from libdebug.debugger.internal_debugger_instance_manager import extend_internal_debugger, provide_internal_debugger
 from libdebug.liblog import liblog
 
+if TYPE_CHECKING:
+    from libdebug.debugger.internal_debugger import InternalDebugger
 
 class MemoryMapList(list[MemoryMap]):
     """A list of memory maps of the target process."""
 
-    def __init__(self: MemoryMapList, memory_maps: list[MemoryMap]) -> None:
+    def __init__(
+        self: MemoryMapList,
+        memory_maps: list[MemoryMap],
+        internal_debugger: InternalDebugger,
+    ) -> None:
         """Initializes the MemoryMapList."""
         super().__init__(memory_maps)
-        self._internal_debugger = provide_internal_debugger(self)
+        self._internal_debugger = internal_debugger
 
     def _search_by_address(self: MemoryMapList, address: int) -> list[MemoryMap]:
         for vmap in self:
@@ -63,8 +70,7 @@ class MemoryMapList(list[MemoryMap]):
         else:
             raise TypeError("The value must be an integer or a string.")
 
-        with extend_internal_debugger(self._internal_debugger):
-            return MemoryMapList(filtered_maps)
+        return MemoryMapList(filtered_maps, self._internal_debugger)
 
     def __hash__(self) -> int:
         """Return the hash of the memory map list."""