--- a/libdebug/data/registers.py
+++ b/libdebug/data/registers.py
@@ -1,20 +1,28 @@
 #
 # This file is part of libdebug Python library (https://github.com/libdebug/libdebug).
-# Copyright (c) 2024 Gabriele Digregorio, Roberto Alessandro Bertolini. All rights reserved.
+# Copyright (c) 2024-2025 Gabriele Digregorio, Roberto Alessandro Bertolini. All rights reserved.
 # Licensed under the MIT license. See LICENSE file in the project root for details.
 #
 
 from __future__ import annotations
 
-from libdebug.debugger.internal_debugger_instance_manager import get_global_internal_debugger
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from libdebug.debugger.internal_debugger import InternalDebugger
 
 
 class Registers:
     """Abtract class that holds the state of the architectural-dependent registers of a process."""
 
-    def __init__(self: Registers, thread_id: int, generic_regs: list[str]) -> None:
+    def __init__(
+        self: Registers,
+        thread_id: int,
+        generic_regs: list[str],
+        internal_debugger: InternalDebugger,
+    ) -> None:
         """Initializes the Registers object."""
-        self._internal_debugger = get_global_internal_debugger()
+        self._internal_debugger = internal_debugger
         self._thread_id = thread_id
         self._generic_regs = generic_regs