--- a/libdebug/data/signal_catcher.py
+++ b/libdebug/data/signal_catcher.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from libdebug.state.thread_context import ThreadContext
 
 
-@dataclass
+@dataclass(eq=False)
 class SignalCatcher:
     """Catch a signal raised by the target process.
 
@@ -59,11 +59,3 @@ class SignalCatcher:
         """Returns whether the signal catcher has been hit on the given thread context."""
         provide_internal_debugger(self)._ensure_process_stopped()
         return self._enabled and thread_context.signal_number == self.signal_number
-
-    def __hash__(self: SignalCatcher) -> int:
-        """Hash the signal catcher object by its memory address, so that it can be used in sets and dicts correctly."""
-        return hash(id(self))
-
-    def __eq__(self: SignalCatcher, other: object) -> bool:
-        """Check if two catchers are equal."""
-        return id(self) == id(other)