--- a/libdebug/data/syscall_handler.py
+++ b/libdebug/data/syscall_handler.py
@@ -6,7 +6,7 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
 from libdebug.debugger.internal_debugger_instance_manager import provide_internal_debugger
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from libdebug.state.thread_context import ThreadContext
 
 
-@dataclass
+@dataclass(eq=False)
 class SyscallHandler:
     """Handle a syscall executed by the target process.
 
@@ -35,14 +35,14 @@ class SyscallHandler:
     syscall_number: int
     on_enter_user: Callable[[ThreadContext, SyscallHandler], None]
     on_exit_user: Callable[[ThreadContext, SyscallHandler], None]
-    on_enter_pprint: Callable[[ThreadContext, int, Any], None]
-    on_exit_pprint: Callable[[int | tuple[int, int]], None]
+    on_enter_pprint: Callable[[ThreadContext, int, Any], None] = field(repr=False)
+    on_exit_pprint: Callable[[int | tuple[int, int]], None] = field(repr=False)
     recursive: bool = False
     hit_count: int = 0
 
-    _enabled: bool = True
-    _has_entered: bool = False
-    _skip_exit: bool = False
+    _enabled: bool = field(default=True, init=False, repr=True)
+    _has_entered: bool = field(default=False, init=False, repr=False)
+    _skip_exit: bool = field(default=False, init=False, repr=False)
 
     @property
     def enabled(self: SyscallHandler) -> bool:
@@ -82,11 +82,3 @@ class SyscallHandler:
         internal_debugger = provide_internal_debugger(self)
         internal_debugger._ensure_process_stopped()
         return self._enabled and thread_context.syscall_number == self.syscall_number and not self._has_entered
-
-    def __hash__(self: SyscallHandler) -> int:
-        """Hash the syscall handler object by its memory address, so that it can be used in sets and dicts correctly."""
-        return hash(id(self))
-
-    def __eq__(self: SyscallHandler, other: object) -> bool:
-        """Check if two handlers are equal."""
-        return id(self) == id(other)