--- a/libdebug/data/syscall_handler.py
+++ b/libdebug/data/syscall_handler.py
@@ -9,11 +9,10 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
-from libdebug.debugger.internal_debugger_instance_manager import provide_internal_debugger
-
 if TYPE_CHECKING:
     from collections.abc import Callable
 
+    from libdebug.debugger.internal_debugger import InternalDebugger
     from libdebug.state.thread_context import ThreadContext
 
 
@@ -43,17 +42,18 @@ class SyscallHandler:
     _enabled: bool = field(default=True, init=False, repr=True)
     _has_entered: bool = field(default=False, init=False, repr=False)
     _skip_exit: bool = field(default=False, init=False, repr=False)
+    _internal_debugger: InternalDebugger = field(default=None, init=True, repr=False)
 
     @property
     def enabled(self: SyscallHandler) -> bool:
         """Returns whether the syscall handler is enabled or not."""
-        provide_internal_debugger(self)._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         return self._enabled
 
     @enabled.setter
     def enabled(self: SyscallHandler, value: bool) -> None:
         """Sets whether the syscall handler is enabled or not."""
-        provide_internal_debugger(self)._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         self._enabled = value
         self._has_entered = False
 
@@ -67,18 +67,15 @@ class SyscallHandler:
 
     def hit_on(self: SyscallHandler, thread_context: ThreadContext) -> bool:
         """Returns whether the syscall handler has been hit on the given thread context."""
-        internal_debugger = provide_internal_debugger(self)
-        internal_debugger._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         return self._enabled and thread_context.syscall_number == self.syscall_number
 
     def hit_on_enter(self: SyscallHandler, thread_context: ThreadContext) -> bool:
         """Returns whether the syscall handler has been hit during the syscall entry on the given thread context."""
-        internal_debugger = provide_internal_debugger(self)
-        internal_debugger._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         return self._enabled and thread_context.syscall_number == self.syscall_number and self._has_entered
 
     def hit_on_exit(self: SyscallHandler, thread_context: ThreadContext) -> bool:
         """Returns whether the syscall handler has been hit during the syscall exit on the given thread context."""
-        internal_debugger = provide_internal_debugger(self)
-        internal_debugger._ensure_process_stopped()
+        self._internal_debugger._ensure_process_stopped()
         return self._enabled and thread_context.syscall_number == self.syscall_number and not self._has_entered