This is still a draft, somehow this change added a few memory leaks?

I'm wrapping everything in a `weakref` next, don't worry.