@MrIndeciso 2025-08-20
Before:
```
Slowest test durations
----------------------------------------------------------------------
1.124s     test_attach_1 (scripts.atexit_handler_test.AtexitHandlerTest.test_attach_1)
1.114s     test_attach_1 (scripts.atexit_handler_test.AtexitHandlerTest.test_attach_1)
1.113s     test_attach_1 (scripts.atexit_handler_test.AtexitHandlerTest.test_attach_1)

----------------------------------------------------------------------
Ran 26600 tests in 1115.850s

OK (skipped=2400)
```
After:
```
Slowest test durations
----------------------------------------------------------------------
1.049s     test_run_4 (scripts.atexit_handler_test.AtexitHandlerTest.test_run_4)
1.038s     test_run_4 (scripts.atexit_handler_test.AtexitHandlerTest.test_run_4)
1.037s     test_attach_1 (scripts.atexit_handler_test.AtexitHandlerTest.test_attach_1)

----------------------------------------------------------------------
Ran 26600 tests in 963.174s

OK (skipped=2400)
```

I had to patch the main branch to actually make the benchmark complete.
The boost in performance is more than I would have expected, maybe the GC is happier now?